Update
Diagnostics table is updated periodically instead of every time when a new message arrives. Heavy CCD-bus activity can slow down the table update and eventually causing delay. Default refresh rate is 50 milliseconds and it can be changed under the table.

Table blinking is reduced by removing the "update" and "refresh" methods when the table is changed.

Fixed a bug where the ID-bytes of PCM/TCM messages were not saved and therefore not evaluated.