Update
Added a new feature that let's the user send a fixed set of messages (at least two) repeatedly, one after another, to the CCD/SCI-bus. In the GUI a new "mode" has been added where you can list the messages separated by commas or semi-colons.

Memory reading on the CCD-bus using B2-requests has been prepared for action. When appropriate the GUI saves the request/response (B2/F2) messages in the "ROMs" folder but it doesn't parse and convert it to a binary file. More to come in a future update.

In other news the TCM-channel (SCI-bus transmission) on the SCI-bus has been temporarily disabled since it's not used for now anyway. Besides transmission controller mainly talk on the CCD-bus.