Legacy OBD1 SCI-bus support for the custom engine cable (CHRY-6)
SCI-bus communication now complies with legacy standards described here:

https://pcm.neons.org/doku.php?id=sci_tech_info

When the OBD1 SCI engine cable is used to communicate with the engine controller select "Settings / OBD1 SCI engine cable used" and send a "Set SCI-bus" settings packet to apply it in the scanner too.

Furthermore, new CCD-bus message interpretations were added.
Also when a B2 request message is echoed on the CCD-bus the previous F2 response message is automatically removed as to not confuse it with the current request.