Improve memory reader method
Exceptions were raised when the read memory form was opened without initializing a session. In this state whatever bus-related communication took place in the main window the other window raised an exception. Log file names are generated during session initialization. Without them any write to the textboxes were problematic.

Some controllers on the SCI-bus may have problems responding to the "return to low-speed mode" command when they are already in low-speed mode, so this part of the code is disabled.