Fix high-speed mode malfunction and low-speed mode actuator tests
Due to the improper use of the sizeof() function on arrays the scanner behaved weirdly. All known array lengths are typed out explicitly.

The weird behavior of the actuator tests are fixed by transmitting the test byte back to the laptop every 200 ms, instead of the way too fast ~20 ms dictated by the PCM. This way the GUI remains responsive and stopping an actuator test becomes more efficient.