Fix CCD-bus memory reader offset check
CCD-bus memory reader offset check is moved to the response handler.
Up to now this check was performed incorrectly in the request handler. This caused the session to finish right after the last request was sent, regardless its success or failure, leading to missing bytes at the end of the binary file.