Implement PCM EEPROM restore feature
Under the Tools / Write memory menu the "RES" button restores EEPROM values from a backup binary file.

First it reads the current memory value and writes a new value only if it's different.

All requests and responses are handled by the scanner. The code is blocking by nature, so unlike with other read/write tasks the scanner and GUI becomes unresponsive until the task is done (~45 seconds) or an error occurs.